requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
cachelib==0.10.2
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
dnacentersdk==2.0.2
fastjsonschema==2.16.2
Flask==2.2.2
Flask-Login==0.6.2
Flask-Session==0.4.0
frozenlist==1.3.3
future==0.18.2
greenlet==2.0.2
identity==0.2.0
idna==2.10
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
meraki==1.4.3
msal==1.21.0
multidict==6.0.4
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==23.0.0
python-dotenv==0.14.0
requests==2.24.0
requests-toolbelt==0.9.1
SQLAlchemy==2.0.3
typing-extensions==4.5.0
urllib3==1.25.10
webexteamssdk==1.6
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.13.0